Restrict chained inequalities to integers for now

+++ source/rust_verify/tests/integers.rs
@@ -88,8 +88,8 @@ test_verify_one_file! {
 test_verify_one_file! {
     #[test] test_chained verus_code! {
         proof fn test1(n: nat) {
-            assert(0 <= n < n + 1 < n + 2);
-            assert(0 <= n + 1 < n < n + 2); // FAILS
+            assert(0 <= n < n as int + 1 < n as int + 2);
+            assert(0 <= n as int + 1 < n < n as int + 2); // FAILS
         }
     } => Err(err) => assert_one_fails(err)
 }
